Image_Classification/requirements.txt
tensorflow
numpy
pillow
matplotlib
opencv-python